+++ user/src/syscall.rs
@@ -103,7 +103,7 @@ pub fn sys_close(fd: usize) -> isize {
     syscall(SYSCALL_CLOSE, [fd, 0, 0])
 }
 
-pub fn sys_pipe(pipefd: &mut [u32; 2]) -> isize {
+pub fn sys_pipe(pipefd: &mut [usize; 2]) -> isize {
     syscall(SYSCALL_PIPE, [pipefd.as_mut_ptr() as usize, 0, 0])
 }
